feat:network state moved to redux
