+++ app-mobile/src/helper/app.ts
@@ -1,0 +1,7 @@
+
+import { Network } from '@capacitor/network';
+
+export const logCurrentNetworkStatus = async () => {
+    const status = await Network.getStatus();
+    return status;
+};
